pyproject.toml
[tool.poetry]
name = "coursework"
version = "0.1.0"
description = ""
authors = ["EgorZH <egor2001zhukov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
